Moved _core into _C module
