+++ pyproject.toml
@@ -2,6 +2,11 @@
 requires = ["scikit-build-core"]
 build-backend = "scikit_build_core.build"
 
+[tool.scikit-build]
+build-dir = "build/{wheel_tag}"
+sdist.include = ["*.pyi"]
+wheel.py-api = "cp312"
+
 [project]
 name = "caterpillar"
 version = "2.0.0-b0"
