+++ src/caterpillar/_C.pyi
@@ -9,6 +9,7 @@ HOST_ARCH: Arch
 InvalidDefault: InvalidDefaultType
 NATIVE_ENDIAN: Endian
 STRUCT_OPTIONS: set
+S_DISCARD_CONST: Option
 S_DISCARD_UNNAMED: Option
 S_EVAL_ANNOTATIONS: Option
 S_REPLACE_TYPES: Option
@@ -158,8 +159,7 @@ class UnaryExpr:
     def __hash__(self) -> int: ...
 
 class atom:
-    @classmethod
-    def __init__(cls, *args, **kwargs) -> None: ...
+    def __init__(self, *args, **kwargs) -> None: ...
     def __pack__(self, obj, ctx) -> Any: ...
     def __size__(self, ctx) -> Any: ...
     def __type__(self) -> Any: ...
@@ -175,8 +175,7 @@ class catom(atom):
     def __unpack_many__(self, *args, **kwargs): ...
 
 class fieldatom(atom):
-    @classmethod
-    def __init__(cls, *args, **kwargs) -> None: ...
+    def __init__(self, *args, **kwargs) -> None: ...
     def __add__(self, other): ...
     def __floordiv__(self, other): ...
     def __getitem__(self, index): ...
@@ -192,8 +191,7 @@ class fieldatom(atom):
     def __xor__(self, other): ...
 
 class fieldcatom(catom):
-    @classmethod
-    def __init__(cls, *args, **kwargs) -> None: ...
+    def __init__(self, *args, **kwargs) -> None: ...
     def __add__(self, other): ...
     def __floordiv__(self, other): ...
     def __getitem__(self, index): ...
@@ -211,6 +209,12 @@ class fieldcatom(catom):
 class fieldinfo:
     excluded: Incomplete
     field: Incomplete
+    def __init__(self, *args, **kwargs) -> None: ...
+
+class intatom(fieldcatom):
+    bytes: Incomplete
+    little_endian: Incomplete
+    signed: Incomplete
     def __init__(self, *args, **kwargs) -> None: ...
 
 class layer:
