+++ test.py
@@ -1,3 +1,3 @@
-from caterpillar._core import intatom, unpack
+from caterpillar._C import intatom, unpack
 
 print(unpack(b"\x01", intatom(1)))
